日付メモリをGraphChartへコンポーネント化
