+++ src/app/(member)/graph/page.tsx
@@ -1,43 +1,10 @@
 "use client";
 
-import {
-  Bar,
-  ComposedChart,
-  Line,
-  ResponsiveContainer,
-  XAxis,
-  YAxis,
-} from "recharts";
 import { useGraphData } from "./_hooks/useGraphData";
 
 export default function GraphPage() {
   // ===== 取得データ =====
-  const { chartData, range, setRange} = useGraphData();
-  
-
-  const formatTick = (v: string) => {
-    const date = new Date(v);
-    if (range === "7days") {
-      return ["日", "月", "火", "水", "木", "金", "土"][date.getDay()];
-    }
-    if (range === "1month") {
-      const day = date.getDate();
-      if (day === 1 || day === 8 || day === 15 || day === 22 || day === 29) {
-        return `${day}日`;
-      }
-      return "";
-    }
-    if (range === "6month") {
-      if (date.getDate() <= 7) {
-        return `${date.getMonth() + 1}月`;
-      }
-      return "";
-    }
-    if (range === "1year") {
-      return `${date.getMonth() + 1}月`;
-    }
-    return `${date.getFullYear()}年`;
-  };
+  const { setRange } = useGraphData();
 
   return (
     <>
@@ -77,24 +44,6 @@ export default function GraphPage() {
           </button>
           {/*UXの時に<br>→<span>で処理 */}
         </div>
-        <ResponsiveContainer width="100%" height={300}>
-          <ComposedChart data={chartData}>
-            <XAxis
-              dataKey="date"
-              tickFormatter={formatTick}
-              stroke="var(--color-text-3)"
-              interval={0}
-            />
-            <YAxis yAxisId="left" domain={[50, 70]} />
-            <YAxis
-              yAxisId="right"
-              orientation="right"
-              domain={["auto", "auto"]}
-            />
-            <Line yAxisId="left" dataKey="weight"></Line>
-            <Bar yAxisId="right" dataKey="steps" fill="pink"></Bar>
-          </ComposedChart>
-        </ResponsiveContainer>
       </div>
     </>
   );
